# openapi.yaml
openapi: 3.0.3
info:
  title: Messenger API
  version: 1.0.0
  description: REST API мессенджера через gateway-service.
servers:
  - url: http://localhost:8080
    description: Local gateway
tags:
  - name: Auth
  - name: User
  - name: Chat
  - name: Message
  - name: Reaction
  - name: Media
paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Register user
      operationId: register
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegistrationDto'
      responses:
        '201':
          description: User registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JWTAuthentificationDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /auth/login:
    post:
      tags: [Auth]
      summary: Login user
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginDto'
      responses:
        '200':
          description: Tokens issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JWTAuthentificationDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh JWT tokens
      operationId: refreshToken
      parameters:
        - $ref: '#/components/parameters/AuthorizationHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Tokens refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JWTAuthentificationDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /user:
    delete:
      tags: [User]
      summary: Delete current user profile and credentials
      operationId: deleteCurrentUser
      responses:
        '204':
          description: User deleted
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /user/{userId}:
    get:
      tags: [User]
      summary: Get user by id
      operationId: getUserById
      parameters:
        - $ref: '#/components/parameters/UserIdPath'
      responses:
        '200':
          description: User profile
          content:
            application/json: {}
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /user/me:
    get:
      tags: [User]
      summary: Get current user
      operationId: getCurrentUser
      responses:
        '200':
          description: Current user profile
          content:
            application/json: {}
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /user/search:
    get:
      tags: [User]
      summary: Search users by tag
      operationId: searchUsersByTag
      parameters:
        - name: tag
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Matching users
          content:
            application/json: {}
        '400':
          $ref: '#/components/responses/BadRequest'
  /user/edit:
    put:
      tags: [User]
      summary: Edit current user
      operationId: editUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EditUserDto'
      responses:
        '200':
          description: Updated user profile
          content:
            application/json: {}
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /user/edit/avatar:
    put:
      tags: [User]
      summary: Edit current user avatar
      operationId: editUserAvatar
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EditAvatarDto'
      responses:
        '200':
          description: Updated user profile
          content:
            application/json: {}
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /chat:
    post:
      tags: [Chat]
      summary: Create chat
      operationId: createChat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateChatDto'
      responses:
        '201':
          description: Chat created
          content:
            application/json: {}
        '400':
          $ref: '#/components/responses/BadRequest'
    get:
      tags: [Chat]
      summary: Get current user's chats
      operationId: getChats
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Chat slice
          content:
            application/json: {}
        '400':
          $ref: '#/components/responses/BadRequest'
  /chat/users:
    post:
      tags: [Chat]
      summary: Add user to chat
      operationId: addUserInChat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddUserInChatDto'
      responses:
        '201':
          description: Chat participant created
          content:
            application/json: {}
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /chat/{chatId}/name/{newName}:
    put:
      tags: [Chat]
      summary: Change chat name
      operationId: changeChatName
      parameters:
        - $ref: '#/components/parameters/ChatIdPath'
        - name: newName
          in: path
          required: true
          schema:
            type: string
            minLength: 1
      responses:
        '200':
          description: Chat name changed
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /chat/{chatId}/my-name:
    put:
      tags: [Chat]
      summary: Set custom chat name for current user
      operationId: setMyCustomChatName
      parameters:
        - $ref: '#/components/parameters/ChatIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetCustomNameDto'
      responses:
        '200':
          description: Custom name changed
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /chat/{chatId}/avatar:
    put:
      tags: [Chat]
      summary: Update chat avatar
      operationId: updateChatAvatar
      parameters:
        - $ref: '#/components/parameters/ChatIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateChatAvatarDto'
      responses:
        '200':
          description: Chat avatar changed
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /chat/role:
    put:
      tags: [Chat]
      summary: Update user role in chat
      operationId: updateChatRole
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRoleUserDto'
      responses:
        '200':
          description: Role updated
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /chat/{chatId}/users:
    get:
      tags: [Chat]
      summary: Get chat participants
      operationId: getChatUsers
      parameters:
        - $ref: '#/components/parameters/ChatIdPath'
      responses:
        '200':
          description: Chat participants
          content:
            application/json: {}
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /chat/{chatId}:
    delete:
      tags: [Chat]
      summary: Delete chat
      operationId: deleteChat
      parameters:
        - $ref: '#/components/parameters/ChatIdPath'
      responses:
        '204':
          description: Chat deleted
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /chat/leave/{chatId}:
    delete:
      tags: [Chat]
      summary: Leave chat
      operationId: leaveChat
      parameters:
        - $ref: '#/components/parameters/ChatIdPath'
      responses:
        '204':
          description: User left chat
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /message:
    post:
      tags: [Message]
      summary: Create message
      operationId: addMessage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMessage'
      responses:
        '202':
          description: Message accepted and published
          content:
            application/json: {}
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /message/chat/{chatId}:
    get:
      tags: [Message]
      summary: Get chat messages
      operationId: getChatMessages
      parameters:
        - $ref: '#/components/parameters/ChatIdPath'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Message slice
          content:
            application/json: {}
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /message/chat/{chatId}/pins:
    get:
      tags: [Message]
      summary: Get pinned messages
      operationId: getPinnedMessages
      parameters:
        - $ref: '#/components/parameters/ChatIdPath'
      responses:
        '200':
          description: Pinned messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PinMessageDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /message/pin/{messageId}:
    post:
      tags: [Message]
      summary: Pin message
      operationId: pinMessage
      parameters:
        - $ref: '#/components/parameters/MessageIdPath'
      responses:
        '201':
          description: Message pinned
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /message/edit:
    put:
      tags: [Message]
      summary: Edit message
      operationId: editMessageById
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MessageEditDto'
      responses:
        '200':
          description: Updated message DTO
          content:
            application/json: {}
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /message/read/{messageId}:
    put:
      tags: [Message]
      summary: Mark message as read
      operationId: readMessageById
      parameters:
        - $ref: '#/components/parameters/MessageIdPath'
      responses:
        '200':
          description: Message marked as read
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /message/read:
    put:
      tags: [Message]
      summary: Mark messages as read
      operationId: readMessageByList
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MessageReadListDto'
      responses:
        '200':
          description: Messages marked as read
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /message/unpin/{messageId}:
    delete:
      tags: [Message]
      summary: Unpin message
      operationId: unpinMessage
      parameters:
        - $ref: '#/components/parameters/MessageIdPath'
      responses:
        '204':
          description: Message unpinned
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /message/{messageId}:
    delete:
      tags: [Message]
      summary: Delete message
      operationId: deleteMessage
      parameters:
        - $ref: '#/components/parameters/MessageIdPath'
      responses:
        '204':
          description: Message deleted
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /reaction/add:
    post:
      tags: [Reaction]
      summary: Add reaction
      operationId: addReaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReactionDto'
      responses:
        '201':
          description: Reaction added
          content:
            application/json: {}
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /reaction/message/{messageId}:
    get:
      tags: [Reaction]
      summary: Get reactions for message
      operationId: getReactionsByMessageId
      parameters:
        - $ref: '#/components/parameters/MessageIdPath'
      responses:
        '200':
          description: Message reactions
          content:
            application/json: {}
        '400':
          $ref: '#/components/responses/BadRequest'
  /reaction/message/{messageId}/{reactionType}:
    delete:
      tags: [Reaction]
      summary: Delete current user's reaction
      operationId: deleteReactionByMessageId
      parameters:
        - $ref: '#/components/parameters/MessageIdPath'
        - name: reactionType
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Reaction deleted
        '400':
          $ref: '#/components/responses/BadRequest'
  /media/upload:
    post:
      tags: [Media]
      summary: Upload media file
      operationId: uploadMedia
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UploadMediaRequest'
      responses:
        '200':
          description: File uploaded
          content:
            application/json: {}
        '400':
          description: Empty file or invalid request
        '500':
          description: Media service error
          content:
            text/plain: {}
  /media/download/{fileName}:
    get:
      tags: [Media]
      summary: Download media file
      operationId: downloadMedia
      parameters:
        - $ref: '#/components/parameters/FileNamePath'
      responses:
        '200':
          description: File bytes
          content:
            image/jpeg: {}
            image/png: {}
            image/gif: {}
            image/webp: {}
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          description: Media service error
          content:
            text/plain: {}
  /media/proxy:
    get:
      tags: [Media]
      summary: Proxy public media file
      operationId: proxyMedia
      parameters:
        - name: publicUrl
          in: query
          required: true
          schema:
            type: string
            minLength: 1
      responses:
        '200':
          description: File bytes
          content:
            image/jpeg: {}
            image/png: {}
            image/gif: {}
            image/webp: {}
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          description: Media service error
          content:
            text/plain: {}

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    AuthorizationHeader:
      name: Authorization
      in: header
      required: true
      schema:
        type: string
        example: Bearer eyJhbGciOiJIUzI1NiJ9...
    UserIdPath:
      name: userId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
    ChatIdPath:
      name: chatId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
    MessageIdPath:
      name: messageId
      in: path
      required: true
      schema:
        type: string
        minLength: 1
    FileNamePath:
      name: fileName
      in: path
      required: true
      schema:
        type: string
        minLength: 1
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        default: 50
        minimum: 1
    Offset:
      name: offset
      in: query
      required: false
      schema:
        type: integer
        default: 0
        minimum: 0
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json: {}
    Unauthorized:
      description: Authentication failed
      content:
        application/json: {}
    NotFound:
      description: Resource not found
      content:
        application/json: {}
    Conflict:
      description: Request conflicts with existing data
      content:
        application/json: {}
  schemas:
    JWTAuthentificationDto:
      type: object
      required: [token, refreshToken]
      properties:
        token:
          type: string
          minLength: 1
          maxLength: 1024
          example: eyJhbGciOiJIUzI1NiJ9.access
        refreshToken:
          type: string
          minLength: 1
          maxLength: 1024
          example: eyJhbGciOiJIUzI1NiJ9.refresh
    UserRegistrationDto:
      type: object
      required: [email, password, name, tag]
      properties:
        email:
          type: string
          format: email
          maxLength: 254
          example: user@example.com
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 72
          example: strongPassword123
        name:
          type: string
          minLength: 4
          maxLength: 50
          example: Ivan Ivanov
        tag:
          type: string
          minLength: 3
          maxLength: 15
          example: ivan
        description:
          type: string
          maxLength: 50
          example: Backend developer
    LoginDto:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          maxLength: 254
          example: user@example.com
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 72
          example: strongPassword123
    RefreshTokenRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
          minLength: 1
          maxLength: 1024
          example: eyJhbGciOiJIUzI1NiJ9.refresh
    EditUserDto:
      type: object
      required: [name, tag]
      properties:
        name:
          type: string
          minLength: 4
          maxLength: 50
          example: Ivan Ivanov
        tag:
          type: string
          minLength: 1
          maxLength: 15
          example: ivan
        description:
          type: string
          maxLength: 50
          example: Backend developer
    EditAvatarDto:
      type: object
      properties:
        newUrl:
          type: string
          example: https://cdn.example.com/avatar.jpg
    CreateChatDto:
      type: object
      required: [name, chatType, participantIds]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 63
          example: Project chat
        chatType:
          type: string
          enum: [PRIVATE, GROUP]
          example: GROUP
        participantIds:
          type: array
          items:
            type: integer
            format: int64
          example: [2, 3]
    AddUserInChatDto:
      type: object
      required: [userId, role, chatId]
      properties:
        userId:
          type: integer
          format: int64
          minimum: 1
          example: 2
        role:
          type: string
          enum: [OWNER, ADMIN, MEMBER]
          example: MEMBER
        chatId:
          type: integer
          format: int64
          minimum: 1
          example: 10
    SetCustomNameDto:
      type: object
      properties:
        name:
          type: string
          example: Work chat
    UpdateChatAvatarDto:
      type: object
      properties:
        avatarUrl:
          type: string
          example: https://cdn.example.com/chat-avatar.jpg
    UpdateRoleUserDto:
      type: object
      required: [userId, chatId, role]
      properties:
        userId:
          type: integer
          format: int64
          minimum: 1
          example: 2
        chatId:
          type: integer
          format: int64
          minimum: 1
          example: 10
        role:
          type: string
          enum: [OWNER, ADMIN, MEMBER]
          example: ADMIN
    CreateMessage:
      type: object
      required: [chatId]
      properties:
        chatId:
          type: integer
          format: int64
          minimum: 1
          example: 10
        content:
          type: string
          example: Hello
        photoLinks:
          type: array
          items:
            type: string
          example: [https://cdn.example.com/photo.jpg]
        repliedMessageId:
          type: string
          pattern: '^$|\S.*'
          example: 65f0c2e5a6f8d45000112233
        forwardedFromMessageId:
          type: string
          pattern: '\S.*'
          example: 65f0c2e5a6f8d45000112234
    MessageEditDto:
      type: object
      required: [id, content]
      properties:
        id:
          type: string
          minLength: 1
          example: 65f0c2e5a6f8d45000112233
        content:
          type: string
          minLength: 1
          example: Updated text
    MessageReadListDto:
      type: object
      required: [ids]
      properties:
        ids:
          type: array
          minItems: 1
          maxItems: 500
          items:
            type: string
            minLength: 1
          example: [65f0c2e5a6f8d45000112233, 65f0c2e5a6f8d45000112234]
    PinMessageDto:
      type: object
      required: [chatId, messageId, messageSendAt, pinnedByUserId]
      properties:
        chatId:
          type: integer
          format: int64
          minimum: 1
          example: 10
        messageId:
          type: string
          minLength: 1
          example: 65f0c2e5a6f8d45000112233
        messageSendAt:
          type: string
          format: date-time
          example: 2026-07-05T10:00:00Z
        pinnedByUserId:
          type: integer
          format: int64
          minimum: 1
          example: 42
    CreateReactionDto:
      type: object
      required: [messageId, reactionType]
      properties:
        messageId:
          type: string
          minLength: 1
          example: 65f0c2e5a6f8d45000112233
        reactionType:
          type: string
          enum: [LIKE, LOVE, WOW, SAD, ANGRY, LAUGH, THINKING, HANDSHAKE, WHITE_HEART, PURPLE_HEART, YELLOW_HEART, GREEN_HEART, BROKEN_HEART, MIND_BLOWN, PRAY, PARTY, HEART_EYES, CRYING, SCREAM, CLAP, COOL]
          example: LIKE
    UploadMediaRequest:
      type: object
      required: [file]
      properties:
        file:
          type: string
          format: binary
security:
  - bearerAuth: []